moving metadata to markket folder
